+++ src/main/java/ru/spaceinvasion/utils/Constants.java
@@ -1,10 +1,10 @@
-package ru.spaceinvasion;
+package ru.spaceinvasion.utils;
 
 public class Constants {
     public static class ApiConstants {
         public static final String BASE_API_PATH = "/v1";
         public static final String USER_API_PATH = BASE_API_PATH + "/user";
-        public static final String SCOREBOARD_API_PATH = BASE_API_PATH + "/leaderboard";
-        public static final int SCOREBOARD_SIZE = 100;
+        public static final String LEADERBOARD_API_PATH = BASE_API_PATH + "/leaderboard";
+        public static final int LEADERBOARD_SIZE = 10;
     }
 }
